A better NAV_BACK logic for ReactNativeClient to save some click when user want to change folder or tag.
When NAV_BACK from folder or tag to the Welcome page, show side menu too.
Because people are mostly going to change to a different folder or tag,
and not just seeing the useless Welcome page at such situation.

People used to press NAV_BACK on Welcome page to exitApp, but when side
menu is shown, NAV_BACK only hide then side menu. This logic should be
better changed to save that click:
When BACK button is pressed on Welcome page while side menu is shown, hide the side menu and also exitApp.
When BACK button is pressed on Welcome page while side menu is not shown, show the side menu.